:bug:  fix: 댓글입력 안 되는 문제 해결
슬님이 새로 짜신 ModalInput에 맞게 변경
