+++ components/Modal/ModalInput/ModalInput.tsx
@@ -10,7 +10,7 @@ import styled from "styled-components";
 interface ModalInputProps {
   label: string;
   $inputType: "댓글" | "제목" | "마감일" | "설명";
-  value?: string;
+  value: string;
   onChange?: (e: React.ChangeEvent<HTMLInputElement | HTMLTextAreaElement>) => void;
 }
 
@@ -23,24 +23,12 @@ interface DateInputProps {
   onChange: (value: string) => void;
 }
 
-const ModalInput = forwardRef<HTMLInputElement | HTMLTextAreaElement, ModalInputProps & { onSubmitComment?: (comment: string) => void }>(
+const ModalInput = forwardRef<HTMLInputElement | HTMLTextAreaElement, ModalInputProps & { onSubmitComment?: () => void }>(
   ({ label, $inputType, value, onChange, onSubmitComment }, ref) => {
-    const [inputValue, setInputValue] = useState("");
     const [selectedDate, setSelectedDate] = useState("");
-
-    const handleInputChange = (e: ChangeEvent<HTMLInputElement | HTMLTextAreaElement>) => {
-      setInputValue(e.target.value);
-    };
 
     const handleDateChange = (dateStr: string) => {
       setSelectedDate(dateStr);
-    };
-
-    const handleCommentSubmit = () => {
-      if (onSubmitComment && inputValue.trim()) {
-        onSubmitComment(inputValue);
-        setInputValue("");
-      }
     };
 
     const renderInput = () => {
@@ -59,7 +47,7 @@ const ModalInput = forwardRef<HTMLInputElement | HTMLTextAreaElement, ModalInput
               />
               {$inputType === "댓글" && (
                 <PositionedButton>
-                  <Button type="modalInput" onClick={handleCommentSubmit} disabled={!inputValue.trim()}>
+                  <Button type="modalInput" onClick={onSubmitComment} disabled={!value.trim()}>
                     입력
                   </Button>
                 </PositionedButton>
